refactor: 优化部分加载逻辑
