+++ framework/sdk-parent/sdk/src/main/resources/commons.properties
@@ -108,4 +108,4 @@ eureka.client.registry-fetch-interval-seconds=10
 # redisson
 spring.redis.redisson.file=file:/opt/metersphere/conf/redisson.yml
 spring.session.redis.repository-type=indexed
-spring.messages.default-locale=zh_CN
+spring.messages.default-locale=
